dropinng unused created_by column

+++ src/main/java/com/livinggoodsbackend/livinggoodsbackend/dto/CreateCommodityUnitRequest.java
@@ -28,7 +28,7 @@ public class CreateCommodityUnitRequest {
 
     private Long linkFacilityId;
     
-    private Long createdBy;
+    private Integer createdBy;
     @NotNull(message = "Total CHPs counted is required")
     private Integer totalCHPsCounted;
 }
